requirements.txt
equinix-metal>=0.2.2
ansible-specdoc>=0.0.13
pydantic==1.10.2
